+++ ui/navigation-menu/CHANGELOG.md
@@ -3,13 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.7.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.7.0...v1.7.1) (2023-06-20)
+## [1.8.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.8.0...v1.8.1) (2023-06-28)
 
 **Note:** Version bump only for package @washingtonpost/wpds-navigation-menu
 
 
 
 
+
+# [1.8.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.7.1...v1.8.0) (2023-06-28)
+
+
+### Features
+
+* new InputSearch component ([#428](https://github.com/washingtonpost/wpds-ui-kit/issues/428)) ([a84c713](https://github.com/washingtonpost/wpds-ui-kit/commit/a84c71392d87ac6d21bc1d1945be814e178ca890))
+
+
+
+
+
+## [1.7.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.7.0...v1.7.1) (2023-06-20)
+
+**Note:** Version bump only for package @washingtonpost/wpds-navigation-menu
 
 # [1.7.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.6.2...v1.7.0) (2023-06-06)